ui(mosaic): Modify grid layout to show preview
The preview of the state was being pushed down because it was starting on
row 2. Try and let the browser handle more of the packing and move the editor as the
second div so it's rendered first. This seems to lead to a layout like

[ docs ]   [ editor]

[ state ]

When viewed at smaller screen resolution, then everything is rendered in a single column.